Put character entities and gemoji into their own files
